+++ .idea/compiler.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="VcsDirectoryMappings">
-    <mapping directory="" vcs="Git" />
+  <component name="CompilerConfiguration">
+    <bytecodeTargetLevel target="21" />
   </component>
 </project>
