+++ app/src/main/java/com/example/figma_replicate/ui/component/ProfileView.kt
@@ -1,0 +1,77 @@
+package com.example.figma_replicate.ui.component
+import android.widget.Toast
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import java.time.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.foundation.layout.*
+import androidx.compose.ui.unit.dp
+import androidx.compose.material3.Icon
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Notifications
+
+import androidx.compose.ui.graphics.Color
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+
+import androidx.compose.foundation.lazy.grid.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+
+@Composable
+fun ProfileView() {
+    val context = LocalContext.current
+
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.spacedBy(40.dp)
+    ) {
+        Spacer(modifier = Modifier.width(10.dp))
+
+        Box(
+            modifier = Modifier
+                .size(50.dp)
+                .clip(CircleShape)
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.profile),
+                contentDescription = "profile picture",
+                contentScale = ContentScale.Crop,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+
+        Column(verticalArrangement = Arrangement.spacedBy(5.dp)) {
+            Text(text = "John Doe", fontWeight = FontWeight.Bold)
+            Text(text = "Lead UI/UX developer")
+        }
+
+        Column {
+            Spacer(modifier = Modifier.height(10.dp))
+            Icon(
+                imageVector = Icons.Filled.Notifications,
+                contentDescription = "Notification icon",
+                tint = Color.Black,
+                modifier = Modifier
+                    .size(25.dp)
+                    .clickable {
+                        Toast.makeText(
+                            context,
+                            "Clicked notification button",
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    }
+            )
+        }
+    }
+}
